1. changed the size of the following variables in the config structure: a) disk_operations_map b) ram_operations_map c) ram_operations_reduce d) disk_operations_reduce from the unsigned long int to the unsigned long long int
The reason for change is that values specified in the config input file usually are greater than the unsigned long int maximum.

+++ MigrationProject/include/common.h
@@ -75,12 +75,12 @@ struct config_s
 	unsigned long int vm_per_host;
 
 	double cpu_flops_map;
-	unsigned long int ram_operations_map;
-	unsigned long int disk_operations_map;
+	unsigned long long int ram_operations_map;
+	unsigned long long int disk_operations_map;
 	unsigned long long int map_output_bytes;
 	double cpu_flops_reduce;
-	unsigned long int ram_operations_reduce;
-	unsigned long int disk_operations_reduce;
+	unsigned long long int ram_operations_reduce;
+	unsigned long long int disk_operations_reduce;
 };
 
 typedef struct config_s* config_t;
